Implement codec from Uint8List.
